+++ OkonkwoOandaV20/OkonkwoOandaV20/TradeLibrary/DataTypes/Instrument/CandleStickData.cs
@@ -2,9 +2,9 @@
 {
    public class CandleStickData
    {
-      public double o { get; set; }
-      public double h { get; set; }
-      public double l { get; set; }
-      public double c { get; set; }
+      public decimal o { get; set; }
+      public decimal h { get; set; }
+      public decimal l { get; set; }
+      public decimal c { get; set; }
    }
 }
