+++ OkonkwoOandaV20/OkonkwoOandaV20/TradeLibrary/DataTypes/Instrument/Instrument.cs
@@ -8,11 +8,11 @@
       public int pipLocation { get; set; }
       public int displayPrecision { get; set; }
       public int tradeUnitsPrecision { get; set; }
-      public double minimumTradeSize { get; set; }
-      public double maximumTrailingStopDistance { get; set; }
-      public double minimumTrailingStopDistance { get; set; }
-      public double maximumPositionSize { get; set; }
-      public double maximumOrderUnits { get; set; }
-      public double marginRate { get; set; }
+      public decimal minimumTradeSize { get; set; }
+      public decimal maximumTrailingStopDistance { get; set; }
+      public decimal minimumTrailingStopDistance { get; set; }
+      public decimal maximumPositionSize { get; set; }
+      public decimal maximumOrderUnits { get; set; }
+      public decimal marginRate { get; set; }
    }
 }
