+++ OkonkwoOandaV20/OkonkwoOandaV20/TradeLibrary/DataTypes/Transaction/DailyFinancingTransaction.cs
@@ -4,8 +4,8 @@ namespace OkonkwoOandaV20.TradeLibrary.DataTypes.Transaction
 {
    public class DailyFinancingTransaction : Transaction
    {
-      public double financing { get; set; }
-      public double accountBalance { get; set; }
+      public decimal financing { get; set; }
+      public decimal accountBalance { get; set; }
       public string accountFinancingMode { get; set; }
       public PositionFinancing positionFinancing { get; set; }
    }
