+++ OkonkwoOandaV20/OkonkwoOandaV20/TradeLibrary/DataTypes/Transaction/TrailingStopLossOrderRejectTransaction.cs
@@ -4,7 +4,7 @@
    {
       public long tradeID { get; set; }
       public string clientTradeID { get; set; }
-      public double distance { get; set; }
+      public decimal distance { get; set; }
       public string timeInForce { get; set; }
       public string gtdTime { get; set; }
       public string triggerCondition { get; set; }
